feat: 점수 높은 진영에 대한 승리 출력 구현

+++ src/main/java/janggi/domain/SideScore.java
@@ -1,4 +1,4 @@
 package janggi.domain;
 
-public record SideScore(int han, int cho) {
+public record SideScore(double han, double cho) {
 }
